@@ -4,6 +4,7 @@ import (
 	"log"
 
 	intentoperatorsregistry "github.com/StripChain/strip-node/intentOperatorsRegistry"
+	db "github.com/StripChain/strip-node/libs/database"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 )
 
@@ -25,7 +26,7 @@ func StartSequencer(
 
 	HeliusApiKey = heliusApiKey
 
-	intents, err := GetIntentsWithStatus(INTENT_STATUS_PROCESSING)
+	intents, err := db.GetIntentsWithStatus(db.INTENT_STATUS_PROCESSING)
 	if err != nil {
 		log.Fatal(err)
 	}
